+++ rust/loqui_client/src/lib.rs
@@ -6,6 +6,7 @@ extern crate log;
 mod client;
 mod config;
 mod connection_handler;
+mod waiter;
 
 pub use client::Client;
 pub use config::Config;
